refactor: move decorator to their own module
