+++ sepal_ui/scripts/utils.py
@@ -5,8 +5,6 @@ import re
 import string
 import warnings
 from configparser import ConfigParser
-from functools import wraps
-from itertools import product
 from pathlib import Path
 from urllib.parse import urlparse
 
@@ -19,8 +17,8 @@ from unidecode import unidecode
 import sepal_ui
 from sepal_ui.conf import config, config_file
 from sepal_ui.message import ms
-
-from .warning import SepalWarning
+from sepal_ui.scripts import decorator as sd
+from sepal_ui.scripts.warning import SepalWarning
 
 
 def hide_component(widget):
@@ -169,132 +167,6 @@ def init_ee():
     return
 
 
-def catch_errors(alert, debug=False):
-    """
-    Decorator to execute try/except sentence
-    and catch errors in the alert message.
-    If debug is True then the error is raised anyway
-
-    Params:
-        alert (sw.Alert): Alert to display errors
-        debug (bool): Wether to raise the error or not, default to false
-    """
-
-    def decorator_alert_error(func):
-        @wraps(func)
-        def wrapper_alert_error(*args, **kwargs):
-            value = None
-            try:
-                value = func(*args, **kwargs)
-            except Exception as e:
-                alert.add_msg(f"{e}", type_="error")
-                if debug:
-                    raise e
-            return value
-
-        return wrapper_alert_error
-
-    return decorator_alert_error
-
-
-def need_ee(func):
-    """
-    Decorator to execute check if the object require EE binding.
-    Trigger an exception if the connection is not possible.
-
-    Params:
-        func (obj): the object on which the decorator is applied
-    """
-
-    @wraps(func)
-    def wrapper_ee(*args, **kwargs):
-
-        # try to connect to ee
-        try:
-            init_ee()
-        except Exception:
-            raise Exception("This function needs an Earth Engine authentication")
-
-        return func(*args, **kwargs)
-
-    return wrapper_ee
-
-
-def loading_button(alert=None, button=None, debug=False):
-    """
-    Decorator to execute try/except sentence and toggle loading button object.
-    Designed to work within the Tile object, or any object that have a self.btn and self.alert set.
-
-    Params:
-        button (sw.Btn, optional): Toggled button
-        alert (sw.Alert, optional): the alert to display the error message
-        debug (bool, optional): wether or not the exception should stop the execution. default to False
-    """
-
-    def decorator_loading(func):
-        @wraps(func)
-        def wrapper_loading(self, *args, **kwargs):
-
-            # set btn and alert
-            # Change name of variable to assign it again in this scope
-            button_ = self.btn if not button else button
-            alert_ = self.alert if not alert else alert
-
-            # Clean previous loaded messages in alert
-            alert_.reset()
-
-            button_.toggle_loading()  # Start loading
-            value = None
-            try:
-                # Catch warnings in the process function
-                with warnings.catch_warnings(record=True) as w_list:
-                    value = func(self, *args, **kwargs)
-
-                # Check if there are warnings in the function and append them
-                # Use append msg as several warnings could be triggered
-                if w_list:
-
-                    # split the warning list
-                    w_list_sepal = [
-                        w for w in w_list if isinstance(w.message, SepalWarning)
-                    ]
-
-                    # display the sepal one
-                    ms_list = [
-                        f"{w.category.__name__}: {w.message.args[0]}"
-                        for w in w_list_sepal
-                    ]
-                    [alert_.append_msg(ms, type_="warning") for ms in ms_list]
-
-                    # only display them in the console if debug mode
-                    if debug:
-
-                        def custom_showwarning(w):
-                            return warnings.showwarning(
-                                message=w.message,
-                                category=w.category,
-                                filename=w.filename,
-                                lineno=w.lineno,
-                                line=w.line,
-                            )
-
-                        [custom_showwarning(w) for w in w_list]
-
-            except Exception as e:
-                alert_.add_msg(f"{e}", "error")
-                if debug:
-                    button_.toggle_loading()  # Stop loading button if there is an error
-                    raise e
-
-            button_.toggle_loading()  # Stop loading button
-
-            return value
-
-        return wrapper_loading
-
-    return decorator_loading
-
-
 def normalize_str(msg, folder=True):
     """
     Normalize an str to make it compatible with file naming (no spaces, special chars ...etc)
@@ -356,105 +228,6 @@ def to_colors(in_color, out_type="hex"):
     return transform(out_color)
 
 
-def switch(*params, debug=True, on_widgets=[], targets=[]):
-    """
-    Decorator to switch the state of input boolean parameters on class widgets or the
-    class itself. If on_widgets is defined, it will switch the state of every widget
-    parameter, otherwise it will change the state of the class (self). You can also set
-    two decorators on the same function, one could affect the class and other the widgets.
-
-    Args:
-        *params (str): any boolean parameter of a SepalWidget.
-        debug (bool): Whether trigger or not an Exception if the decorated function fails.
-        on_widgets (list(widget_names,)|optional): List of widget names into the class
-        targets (list(bool,)|optional); list of the target value (value taht will be set on switch. default to the inverse of the current state.
-
-    """
-
-    def decorator_switch(func):
-        @wraps(func)
-        def wrapper_switch(self, *args, **kwargs):
-
-            widgets_len = len(on_widgets)
-            targets_len = len(targets)
-
-            # sanity check on targets and on_widgets
-            if widgets_len and targets_len:
-                if widgets_len != targets_len:
-                    raise IndexError(
-                        f'the length of "on_widgets" ({widgets_len}) is different from the length of "targets" ({targets_len})'
-                    )
-
-            # create the list of target values based on the target list
-            # or the initial values of the widgets params
-            # The first one is taken as reference
-            if not targets_len:
-                w = getattr(self, on_widgets[0]) if widgets_len else self
-                targets_ = [bool(getattr(w, p)) for p in params]
-            else:
-                targets_ = targets
-
-            if widgets_len:
-
-                # Verify that the input elements are strings
-                wrong_types = [
-                    (w, type(w)) for w in on_widgets if not isinstance(w, str)
-                ]
-
-                if len(wrong_types):
-                    errors = [
-                        f"Received:{w_type} for widget: {w}."
-                        for w, w_type in wrong_types
-                    ]
-
-                    raise TypeError(
-                        f"All on_widgets list elements has to be strings. [{' '.join(errors)}]"
-                    )
-
-                missing_widgets = [w for w in on_widgets if not hasattr(self, w)]
-
-                if missing_widgets:
-                    raise Exception(
-                        f"The provided {missing_widgets} widget(s) does not exist in the current class"
-                    )
-
-                def w_assign(bool_targets):
-
-                    params_targets = [
-                        (p, bool_targets[i]) for i, p in enumerate(params)
-                    ]
-
-                    for (w_name, p_t) in product(on_widgets, params_targets):
-                        param, target = p_t
-                        widget = getattr(self, w_name)
-                        setattr(widget, param, target)
-
-            else:
-
-                def w_assign(bool_targets):
-                    for i, p in enumerate(params):
-                        setattr(self, p, bool_targets[i])
-
-            # assgn the parameters to the target inverse
-            w_assign([not t for t in targets_])
-
-            # execute the function and catch errors
-            try:
-                func(self, *args, **kwargs)
-
-            except Exception as e:
-                if debug:
-                    w_assign(targets_)
-                    raise e
-
-            # reassign the parameters to the targets
-            w_assign(targets_)
-
-        return wrapper_switch
-
-    return decorator_switch
-
-
 def next_string(string):
     """
     Create a string followed by an underscore and a consecutive number
@@ -505,7 +278,6 @@ def set_config(key, value, section="sepal-ui"):
 @deprecated(
     version="2.9.1", reason="This function will be removed in favor of set_config()"
 )
-@versionadded(version="2.7.0")
 def set_config_locale(locale):
     """
     Set the provided local in the sepal-ui config file
@@ -536,7 +308,6 @@ def set_config_locale(locale):
 @deprecated(
     version="2.9.1", reason="This function will be removed in favor of set_config()"
 )
-@versionadded(version="2.7.0")
 def set_config_theme(theme):
     """
     Set the provided theme in the sepal-ui config file
@@ -661,3 +432,15 @@ def check_input(input_, msg=ms.utils.check_input.error):
         raise ValueError(msg)
 
     return init
+
+
+################################################################################
+# the soon to be deprecated decorators
+#
+
+# fmt: off
+catch_errors = deprecated(version='3.0', reason="use sepal_ui.scripts.decorator.catch_errors instead")(sd.catch_errors)
+need_ee = deprecated(version='3.0', reason="use sepal_ui.scripts.decorator.need_ee instead")(sd.need_ee)
+loading_button = deprecated(version='3.0', reason="use sepal_ui.scripts.decorator.need_ee instead")(sd.loading_button)
+switch = deprecated(version='3.0', reason="use sepal_ui.scripts.decorator.switch instead")(sd.switch)
+# fmt: on
